move WalletConnect to LegacyWalletConnect
