+++ src/screens/LegacyWalletConnect/WalletConnectActiveConnections.js
