+++ src/screens/LegacyWalletConnect/WalletConnectCallRequest.js
