+++ src/screens/LegacyWalletConnect/WalletConnectCallRequestList.js
