+++ src/screens/LegacyWalletConnect/WalletConnectConnectorRequest.js
