+++ src/screens/LegacyWalletConnect/WalletConnectDappsPromoCard.js
