+++ src/screens/LegacyWalletConnect/WalletConnectDashboard.js
