+++ src/screens/LegacyWalletConnect/WalletConnectPinConfirm.js
