+++ src/screens/LegacyWalletConnect/WalletConnectQRCodeScanButton.js
