+++ src/screens/LegacyWalletConnect/package.json
